requirements.txt
Flask
flask-cors
pandas
openpyxl
Pillow